% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensitivity.R
\name{get_fnr}
\alias{get_fnr}
\title{Get test false negative rate x days from crossing the threshold of detection}
\usage{
get_fnr(time, shape = 2.33, rate = 0.24, max_sensitivity = 0.99)
}
\arguments{
\item{time}{Numeric. Number of days from crossing the threshold of detection}

\item{shape}{Numeric. Shape of the gamma distribution used to estimate
test sensitivity. Default is 2.33}

\item{rate}{Numeric. Rate of the gamma distribution used to estimate test
sensitivity. Default is 0.24}

\item{max_sensitivity}{Numeric. Value between 0 and 1 indicating the maximum
sensitivity. Default is 0.99.}
}
\description{
Get test false negative rate x days from crossing the threshold of detection
}
\examples{
get_fnr(1)
}
