% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob_missed_infection.R
\name{get_prob_missed_infection}
\alias{get_prob_missed_infection}
\title{Calculate the probability of a missed infection given a test + quarantine
strategy}
\usage{
get_prob_missed_infection(
  test_time,
  additional_quarantine_time = 0,
  tolerance = 0.01,
  shape_exposure_to_threshold = 1.98,
  shape_threshold_to_symptoms = 2.16,
  rate = 0.72
)
}
\arguments{
\item{test_time}{Numeric. Date post-exposure test sample is collected}

\item{additional_quarantine_time}{Numeric. Additional days quarantined after
test sample is collected. Default is 0}

\item{tolerance}{Numeric. Tolerance. Default is 0.01}

\item{shape_exposure_to_threshold}{Numeric. Shape of the gamma distribution
that describes the time from exposure to crossing the threshold of
detection. Default is 1.98}

\item{shape_threshold_to_symptoms}{Numeric. Shape of the gamma distribution
that describes the time from crossing the threshold of
detection to symptom onset. Default is 2.16}

\item{rate}{Numeric. Rate of the gamma distributions that describe the time
from exposure to crossing the threshold of detection and the time from
crossing the threshold of detection to symptom onset. Default is 0.72.}
}
\description{
Calculate the probability of a missed infection given a test + quarantine
strategy
}
\examples{
## test sample collected 5 days post-exposure, quarantine until day 5
get_prob_missed_infection(5)

## test sample collected 7 days post-exposure, quarantine until day 9
get_prob_missed_infection(7, 2)
}
