% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensitivity.R
\name{get_sensitivity}
\alias{get_sensitivity}
\title{Get test sensitivity x days from crossing the threshold of detection}
\usage{
get_sensitivity(time, shape = 2.33, rate = 0.24)
}
\arguments{
\item{time}{Numeric. Number of days from crossing the threshold of detection}

\item{shape}{Numeric. Shape of the gamma distribution used to estimate
test sensitivity. Default is 2.33}

\item{rate}{Numeric. Rate of the gamma distribution used to estimate test
sensitivity. Default is 0.24}
}
\description{
Get test sensitivity x days from crossing the threshold of detection
}
\examples{
get_sensitivity(1)
}
