% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shenzhen_estimate.R
\name{get_shenzhen_estimate}
\alias{get_shenzhen_estimate}
\title{Extract a single estimate of the Shenzhen sensitivity based on the Bayesian
logistic regression}
\usage{
get_shenzhen_estimate(x)
}
\arguments{
\item{x}{Numeric. Days since symptom onset}
}
\description{
Extract a single estimate of the Shenzhen sensitivity based on the Bayesian
logistic regression
}
\examples{
get_shenzhen_estimate(0)
}
